Remove uninformative coverage stuff
Though it would be nice to integrate my build/test process with
https://coveralls.io/, the coverage I can get out of numba.jitted
functions is uninformative.  Basically, the coverage code thinks that
jitted functions have not been run.  I've asked a question about this
<http://stackoverflow.com/questions/26875191/analyzing-coverage-of-numba-wrapped-functions>,
on stackoverflow but I don't expect a response.  In principle, it
seems like it might be possible using llvm-cov.  Thus, currently, I
don't see a way of using this sort of thing.