Add jitted functions to compute LM and LMpM arrays
The LMpM arrays (arrays of [ell,mp,m] for all the values between some
ell_min and ell_max) can actually be quite time-consuming to compute,
especially given that the tests need to compute these many times.
This commit adds a couple new functions to do this quickly.  There is
a related LM function; LM is much easier to calculate, so the gains
are minimal, but for uniformity, let's include it.