Add inverse function for ethbar_NP and tests; also jit mode conversion functions
A bunch of these mode-conversion functions could benefit from a simple jit decorator.  They won't be totally jitted, but the loops can be compiled.

The inverse function for ethbar_NP is useful for when I want a supertranslation that will give me a specified contribution to the strain.

This is also easy to test, by simply round-tripping some random data for spins ranging from -5 to 5, and ensuring that everything is the same, except for the special cases where information is necessarily lost.