Fix sign of D matrix elements when Ra = 0
The final simplification in equation (3) of the "Wigner D matrices"
page introduced an error by exchanging m' for m when the Kronecker
delta indicates that m' should be exchanged for -m. This fixes the
error and brings the documentation into agreement with the behavior
of `spherical_functions.Wigner_D_element`.